front/requirements.txt
anyio==3.6.2
asn1crypto==0.24.0
attrs==21.4.0
bcrypt==4.0.1
certifi==2018.8.24
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
cloud-init==19.1
configobj==5.0.6
cryptography==2.6.1
docker==6.1.2
ecdsa==0.18.0
entrypoints==0.3
fastapi==0.95.2
Flask==2.2.5
Flask-Cors==3.0.10
greenlet==2.0.2
h11==0.14.0
httplib2==0.11.3
idna==3.4
importlib-metadata==4.11.1
importlib-resources==5.4.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.2
jsonschema==4.4.0
keyring==17.1.1
keyrings.alt==3.1.1
MarkupSafe==2.1.2
oauthlib==3.2.0
packaging==23.1
passlib==1.7.4
psycopg2==2.7.7
pyasn1==0.5.0
pycrypto==2.6.1
pycurl==7.43.0.2
pydantic==1.10.7
PyGObject==3.30.4
pyrsistent==0.18.1
PySimpleSOAP==1.16.2
python-apt==1.8.4.3
python-debian==0.1.35
python-debianbts==2.8.2
python-dotenv==0.21.1
python-jose==3.3.0
python-multipart==0.0.6
pyxdg==0.25
PyYAML==6.0
reportbug===7.5.3-deb10u1
requests==2.30.0
rsa==4.9
SecretStorage==2.3.1
six==1.12.0
sniffio==1.3.0
SQLAlchemy==2.0.13
starlette==0.27.0
typing_extensions==4.5.0
urllib3==2.0.2
uvicorn==0.22.0
websocket-client==1.5.1
Werkzeug==2.2.3
zipp==3.7.0
